move session handler traits to frame-support
