+++ frame/session/src/mock.rs
@@ -119,7 +119,7 @@ impl ShouldEndSession<u64> for TestShouldEndSession {
 }
 
 pub struct TestSessionHandler;
-impl SessionHandler<u64> for TestSessionHandler {
+impl frame_support::traits::SessionHandler<u64> for TestSessionHandler {
 	const KEY_TYPE_IDS: &'static [sp_runtime::KeyTypeId] = &[UintAuthorityId::ID];
 	fn on_genesis_session<T: OpaqueKeys>(_validators: &[(u64, T)]) {}
 	fn on_new_session<T: OpaqueKeys>(
